requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.1
pillow==10.4.0
PyMySQL==1.1.1
